backend/requirements_py313.txt
# Python 3.13兼容版本的依赖文件
# Web框架 - 升级到支持Python 3.13的版本
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
starlette>=0.35.0

# Pydantic - 必须使用支持Python 3.13的版本
pydantic>=2.8.0
pydantic-settings>=2.4.0

# 数据库
SQLAlchemy>=2.0.25
alembic>=1.13.0

# 认证和安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
PyJWT>=2.8.0
bcrypt>=4.2.0

# AI和ML相关 - 使用较新版本以支持Python 3.13
openai>=1.30.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
transformers>=4.40.0
torch>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.4.0
scipy>=1.12.0

# 音频处理
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1
SpeechRecognition>=3.10.0
edge-tts>=6.1.0
gTTS>=2.4.0
pyttsx3>=2.90
ffmpeg-python>=0.2.0

# 其他AI服务
dashscope>=1.20.0
modelscope>=1.15.0

# HTTP和API
httpx>=0.27.0
requests>=2.31.0
aiohttp>=3.9.0

# 数据处理
jsonschema>=4.20.0

# 工具库
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0
loguru>=0.7.0
tenacity>=8.0.0

# 测试
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
pytest-timeout>=2.3.0
coverage>=7.0.0

# 开发工具
GitPython>=3.1.40
colorama>=0.4.6

# Ollama相关
ollama>=0.3.0

# 其他依赖
certifi>=2024.2.2
charset-normalizer>=3.3.0
urllib3>=2.2.0
idna>=3.6
packaging>=24.0
setuptools>=69.0.0
wheel>=0.42.0